refactor: move version.ts into src/utils/
Establish src/utils/ as the home for shared utility modules.
paths.ts and config.ts will follow in subsequent PRs.
